Fixed #1337 re: link error with the PLC stuff if G.711 codec is disabled.
This basically is caused by the problem with library order in the Makefile. Since libpjmedia (which contains the PLC code) is specified first before the third party libs (which the codecs reside), the PLC didn't get picked up by the linker. It works if G.711 is enabled, because G.711 lives in libpjmedia.

The real problem is because there is a circular dependency between pjmedia and third party libraries. The real solution would be to break down third party libraries into individual lib and arrange the link order correctly. This would need a bit of work in the Makefile though.

For now, let's just hack pjmedia-test so that it links with the PLC if G.711 is disabled.



git-svn-id: http://svn.pjsip.org/repos/pjproject/branches/1.x@3681 74dad513-b988-da41-8d7b-12977e46ad98